Crate client

A simple TCP client for sending files to a server.

This client connects to a specified TCP server and sends the contents of a file. It uses the log crate for logging errors and information during the execution of the program.

Usage

To run the client, specify the server address, port number, file path, and a placeholder argument:

cargo run -- <address> <port> <file_path> <go_back_n>

Replace <address> with the server’s IP address (e.g., 127.0.0.1), <port> with the desired port number, and <file_path> with the path to the file you want to send.

Functions

handle_frame 🔒

Handles the received frames from the server.

main 🔒

The main function that initializes the client.

receive 🔒

Receives frames from the server and handles them. The function reads from the stream and constructs frames from the received bytes. It then calls handle_frame() to process the frames. If a connection end frame is received, the function returns.

send 🔒

Sends a file to the specified server address.

Type Aliases

SafeCond 🔒

SafeWindow 🔒