Module client::camera

Provides a camera resource to store the view perspective of the user.

Modules

unsafe_proj

Visibiilty guard to avoid inconsistent proj updates.

Structs

Camera

A resource that stores the view perspective of the user.

Constants

GL_RIGHT_DIR

Rightward axis for clip space.

GL_TOP_DIR

Upward axis for clip space.

GL_VIEW_DIR

Forward axis for clip space.

Functions

debug

debug_setup

debug_system

setup_ecs

Sets up legion ECS for this module.