Struct client::camera::Camera

pub struct Camera { /* fields omitted */ }

A resource that stores the view perspective of the user.

Implementations

impl Camera

pub fn focus(&self) -> Position

The point focused by the camera.

projection() shall transform this point to (0.5, 0.5, z), where z = zoom / distance.

pub fn rotation(&self) -> Matrix

The rotation matrix of the camera.

This transforms the real coordinates to the coordinates as seen by the user.

pub fn aspect(&self) -> f64

Canvas width divided by canvas height

pub fn zoom(&self) -> f64

The distance of the focus from the camera.

pub fn distance(&self) -> f64

The rendering distance of the camera.

pub fn fovy(&self) -> f64

The vertical field of view in radians.

impl Camera

pub fn builder() -> CameraBuilder<((), (), (), (), (), (), (), ())>

Create a builder for building Camera. On the builder, call .focus(...), .rotation(...), .aspect(...), .zoom(...), .distance(...), .fovy(...), .proj(...)(optional), .proj_inv(...)(optional) to set the values of the fields. Finally, call .build() to create the instance of Camera.

impl Camera

pub fn set_focus(&mut self, focus: Position)

Sets the point focused by the camera.

The focus point is transformed to (0.5, 0.5, zoom/distance) by the projection matrix.

pub fn set_rotation(&mut self, rotation: Matrix)

Sets the rotation matrix of the camera.

This transforms the real coordinates to the coordinates as seen by the user.

pub fn set_aspect(&mut self, aspect: f64)

Sets the aspect ratio, which is the canvas width divided by the canvas height.

pub fn set_zoom(&mut self, zoom: f64)

Sets the distance of focus from the camera.

pub fn set_distance(&mut self, distance: f64)

Sets the rendering distance of the camera.

pub fn set_fovy(&mut self, fovy: f64)

Sets the vertical Field of View in radians.

pub fn projection(&self) -> Matrix

Transforms real coordinates to unit cube [0, 1]^3

pub fn inv_projection(&self) -> Matrix

Transforms unit cube [0, 1]^2 to real coordinates

pub fn asymptotic_projection(&self) -> LinearMatrix

An “asymptotic” version of the projection matrix, which does not perform translation.

impl Camera

pub fn project_mouse(&self, x: f64, y: f64) -> (Position, Position)

Projects a mouse click from the unit square [0, 1]^2 to real coordinates.

Returns

Returns a tuple (a, b). The mouse clicks on points between points a and b. a is the closer point and b is the further point.