client::conexion::handler_archivo

Struct HandlerArchivo

pub struct HandlerArchivo<R, W>
where
    R: BufRead,
    W: Write,
{
    archivo: String,
    reader_stream: R,
    writer_stream: W,
    buffer: String,
}

Maneja la conexion con un servidor enviandole las operacion que contiene un archivo.

Fields
archivo: String

El nombre del archivo del cual se leeran las operaciones.

reader_stream: R

Canal por el cual se leeran las respuestas del servidor.

writer_stream: W

Canal por el cual se le enviaran las operaciones al servidor.

buffer: String

En donde se almacenaran los mensajes enviados por el servidor.

Implementations

impl<R, W> HandlerArchivo<R, W>
where R: BufRead, W: Write,

pub fn new( archivo: String, reader_stream: R, writer_stream: W, buffer: String, ) -> Self

Crea y devuelve una instancia de HandlerArchivo.

pub fn gestionar(&mut self) -> Result<(), ErrorEnCliente>

Envia cada linea del archivo (que representa una operacion) al servidor y maneja su respuesta.

Una vez que termina de procesar el archivo le pide un resultado final.

En caso de algun error retorna ErrorEnCliente.

fn abrir_archivo(&self) -> Result<BufReader<File>, ErrorEnCliente>

Abre el archivo y lo devuelve dentro de un BufReader para poder manejarlo mas comodamente.

Si ocurre algun error retorna un ErrorEnCliente.

fn enviar_mensaje_servidor( &mut self, mensaje: &str, operacion: Option<String>, ) -> Result<(), ErrorEnCliente>

Envia un mensaje al servidor a traves del writer_stream.

Si ocurre algun error retorna un ErrorEnCliente.

fn leer_y_manejar_respuesta_servidor(&mut self) -> Result<(), ErrorEnCliente>

Lee la respuesta del servidor a traves del reader_stream y se encarga de manejarla.

Si ocurre algun error retorna un ErrorEnCliente.

fn manejar_output_respuesta( &self, respuesta: &str, ) -> Result<(), ErrorEnCliente>

Le muestra al cliente en caso de ser necesario lo que el servidor le envio como respuesta.

Si el servidor le indica al cliente que hubo un error con su peticion se lo indica por stderr.

Si ocurre algun error retorna un ErrorEnCliente.