Function client::receive

pub(crate) async fn receive(
    stream: OwnedReadHalf,
    window: Arc<Mutex<Window>>,
    condition: Arc<Condvar>,
) -> Result<(), ()>

Receives frames from the server and handles them. The function reads from the stream and constructs frames from the received bytes. It then calls handle_frame() to process the frames. If a connection end frame is received, the function returns.

Arguments

stream - The read half of the TCP stream.

window - The window to manage the frames.