Function client::send_file

pub(crate) fn send_file(
    tx: Sender<Vec<u8>>,
    safe_window: SafeWindow,
    condition: SafeCond,
    file_path: String,
) -> JoinHandle<Result<&'static str, &'static str>>

Sends a file to the specified server address.

This function opens the specified file, reads it and sends it over the established TCP connection chunk by chunk. It creates a new frame for each chunk and sends it to the server. The function also handles resending frames in case of a timeout.

Arguments

tx - The sender channel to send the frame to the writer task.

safe_window - The window to manage the frames.

condition - The condition to signal tasks that space is available in the window.

file_path - The path to the file to be sent.

Panics

This function will panic if:

The window mutex cannot be locked.

The condition cannot be waited on.

The frame cannot be pushed to the window.

The frame cannot be sent to the writer task.

Returns

The function returns a JoinHandle corresponding to the async task launched that can be awaited to get the result of the operation.