Struct client::html::Scope

pub struct Scope<COMP> where
    COMP: Component,  {
    parent: Option<Rc<AnyScope>>,
    state: Rc<RefCell<Option<ComponentState<COMP>>>>,
}

A context which allows sending messages to a component.

Fields
parent: Option<Rc<AnyScope>>state: Rc<RefCell<Option<ComponentState<COMP>>>>
Implementations

impl<COMP> Scope<COMP> where
COMP: Component,

pub fn get_parent(&self) -> Option<&AnyScope>

Returns the parent scope

pub fn get_component(&self) -> Option<impl Deref<Target = COMP>>

Returns the linked component if available

pub fn send_message<T>(&self, msg: T) where
T: Into<<COMP as Component>::Message>,

Send a message to the component.

Please be aware that currently this method synchronously schedules a call to the Component interface.

pub fn send_message_batch(
&self,
messages: Vec<<COMP as Component>::Message, Global>
)

Send a batch of messages to the component.

This is useful for reducing re-renders of the components because the messages are handled together and the view function is called only once if needed.

Please be aware that currently this method synchronously schedules calls to the Component interface.

pub fn callback<F, IN, M>(&self, function: F) -> Callback<IN> where
M: Into<<COMP as Component>::Message>,
F: 'static + Fn(IN) -> M,

Creates a Callback which will send a message to the linked component’s update method when invoked.

Please be aware that currently the result of this callback synchronously schedules a call to the Component interface.

pub fn callback_once<F, IN, M>(&self, function: F) -> Callback<IN> where
M: Into<<COMP as Component>::Message>,
F: 'static + FnOnce(IN) -> M,

Creates a Callback from an FnOnce which will send a message to the linked component’s update method when invoked.

Please be aware that currently the result of this callback will synchronously schedule calls to the Component interface.

pub fn batch_callback<F, IN, OUT>(&self, function: F) -> Callback<IN> where
F: 'static + Fn(IN) -> OUT,
OUT: SendAsMessage<COMP>,

Creates a Callback which will send a batch of messages back to the linked component’s update method when invoked.

The callback function’s return type is generic to allow for dealing with both Option and Vec nicely. Option can be used when dealing with a callback that might not need to send an update.

link.batch_callback(|_| vec![Msg::A, Msg::B]);
link.batch_callback(|_| Some(Msg::A));

Please be aware that currently the results of these callbacks will synchronously schedule calls to the Component interface.

pub fn batch_callback_once<F, IN, OUT>(&self, function: F) -> Callback<IN> where
F: 'static + FnOnce(IN) -> OUT,
OUT: SendAsMessage<COMP>,

Creates a Callback from an FnOnce which will send a batch of messages back to the linked component’s update method when invoked.

The callback function’s return type is generic to allow for dealing with both Option and Vec nicely. Option can be used when dealing with a callback that might not need to send an update.

link.batch_callback_once(|_| vec![Msg::A, Msg::B]);
link.batch_callback_once(|_| Some(Msg::A));

Please be aware that currently the results of these callbacks will synchronously schedule calls to the Component interface.