Module client::models

Modules

api_version

contest

invoke_status

participation

problem

run

run_live_status_update

run_patch

run_simple_submit_params

session_token

simple_auth_params

toolchain

user

user_create_params

Structs

ApiVersion

Contest

InvokeStatus

Participation

Problem

Run
Run : Represents a run.

RunLiveStatusUpdate
RunLiveStatusUpdate : Represents Live Status Update Some fields can be missing for various reasons, it is normal that particular object will look like {liveScore: null, currentTest: null, finish: false}. Information in all fields except finish can be inaccurate, incorrect or outdated. You can rely on following: if finish is true, final judging results are available

RunPatch

RunSimpleSubmitParams

SessionToken
SessionToken : Type that represents session You shouldn't do any assumptions about this type representation

SimpleAuthParams

Toolchain

User

UserCreateParams