client
Module networking

Functions

connect_to_server

Connect to a server at the specified address.

convert_to_pixel_data 🔒

Convert a Vec<u8> to a PixelData struct.

process_fragment_task

Process a FragmentTask by generating a fractal image and saving it to the filesystem.

receive_fragment_task

Receive a FragmentTask from the server.

save_fractal_image

Save a fractal image to the filesystem.

send_fragment_result 🔒

Send a FragmentResult to the server after generating a fractal image.

send_request

Send a request to the server.