Trait client::options::ColorMapCount

pub trait ColorMapCount<T> {
    fn is_some(&self) -> bool;

    fn feed(&mut self, comps: T);

    fn compute(&self, comps: T) -> Vector;
}

A general trait backed by ColorMapCounter on different closure types, implemented on Option to support unused filters.

Required methods

fn is_some(&self) -> bool

Proxies Option::is_some.

fn feed(&mut self, comps: T)

Feed the components into the counter, which forwards to the function for type counting.

Do not call after using ColorMapCount::compute.

fn compute(&self, comps: T) -> Vector

Compute the color given the components, using the extrema collected in ColorMapCount::feed.

Only use after calling ColorMapCount::feed on all entities.