Module client::protocol::message

A protocol message is defined as follows:

component	size (bits)	description

msg_type	2	The type (action) of the message

msg_length	6	The length of the entire message

msg_payload	0 - 62 bytes	The payload of the message

checksum	8	The checksum of the message

Note that:

the msg_type and msg_length form the first byte of the message, which is the first two bits of the byte for msg_type and the following six bits for msg_length.

the value of msg_length is the length of the entire message in bytes, as 6 (unsigned) bits can hold value from 0 to 63. Its maximum value is defined as MAX_PROTOCOL_SIZE. So a single protocol message is capable of carrying MAX_PROTOCOL_SIZE - 2 = 61 bytes as payload, this value is denoted as MAX_PAYLOAD_SIZE.

the checksum is not included in the Message struct, it is calculated and appended when serializing the message. Similarly, the process of deserializing will also validate the checksum.

Structs

Message

A protocol message. See the module-level documentation for more details.

Enums

MessageType

The type (action) of a message which is the first two bits of a message.

Constants

MAX_PAYLOAD_SIZE

Maximum size of payload comes from MAX_PROTOCOL_SIZE - 2.

MAX_PROTOCOL_SIZE

Maximum size of a protocol message. It is counted in bytes.

Type Aliases

Byte

A byte is logically equivalent to an 8-bit unsigned integer.