Trait client::render::scene::mesh::AbstractPreparedMesh

pub trait AbstractPreparedMesh {
    fn positions(&self) -> &FloatBuffer;

    fn normals(&self) -> &FloatBuffer;

    fn tex_pos(&self) -> &FloatBuffer;

    fn tex_sprite_number(&self) -> &[usize];

    fn tex_offset(&self) -> &FloatBuffer;

    fn draw(&self, gl: &WebGlRenderingContext);
}

A generic mesh prepared for WebGL rendering.

Required methods

fn positions(&self) -> &FloatBuffer

Buffer storing vertex positions.

Corresponds to a_pos.

fn normals(&self) -> &FloatBuffer

Buffer storing vertex normals.

Corresponds to a_normal.

fn tex_pos(&self) -> &FloatBuffer

Buffer storing texture positions.

Corresponds to a_tex_pos.

fn tex_sprite_number(&self) -> &[usize]

A vector storing the sprite number of each vertex.

This is used with the actual sprite definition to populate tex_offset.

fn tex_offset(&self) -> &FloatBuffer

Buffer transferring sprite-specific texture offsets.

Corresponds to a_tex_offset.

fn draw(&self, gl: &WebGlRenderingContext)

Draws the whole mesh on the canvas.