Module client::render::texture

Sprite sheet definition.

Structs

Atlas

An atlas, which is a single image containing many small sprites.

CubeSprites

Sprites for a cube.

CylinderSprites

Sprites for a cylinder.

Icon

An icon struct ready for HTML use.

Index

The loaded index of an atlas.

Pool

Stores atlas cache.

PreparedTexture

A texture that can be used on WebGL directly.

RectSprite

A rectangular sprite.

Enums

AtlasEnum

ShapeSprites

Sprites for one shape.

Constants

DUMMY_CUBE_SPRITES

DUMMY_RECT_SPRITE

Functions

decompose_value

Decompose the load_textures response into index and bitmap fields and parse the index.

get_bitmap

get_index

load_textures