Module client::render::util

Low-level WebGL wrappers.

Modules

buffer

shader

types

uniform

Structs

AttrLocation

An attribute location to bind a buffer to.

FloatBuffer

A buffer of float values to be passed to a WebGL program.

IndexBuffer

A buffer of index values to be passed to a WebGL program.

UniformLocation

Stores a uniform location with known type.

Enums

BufferUsage

Usage pattern of a buffer.

Traits

Glize

A float-based type that can be lossily converted for WebGL compatibility.

Uniform

A type that can be assigned as a uniform.

Functions

create_program

Creates a WebGL program.

create_shader

Compiles a WebGL shader.