Module client::rpc::errors

Modules

codes

RPC Error codes and error objects

error_codes

RPC Error codes and error objects

Structs

RevertError

Represents a reverted transaction and its output data.

TransactionInputError

Error thrown when both data and input fields are set and not equal.

Enums

EthApiError

Errors that can occur when interacting with the eth_ namespace

RpcInvalidTransactionError

An error due to invalid transaction.

RpcPoolError

A helper error type that’s mainly used to mirror geth Txpool’s error messages

Functions

build_rpc_server_error

call_execution_error

error_object_owned_to_jsonrpc_error

geth_call_execution_error

internal_error

internal_error_msg

internal_rpc_err

Constructs an internal JSON-RPC error.

internal_rpc_err_with_data

Constructs an internal JSON-RPC error with data

invalid_input_rpc_err

invalid_params

invalid_params_check

invalid_params_detail

invalid_params_msg

invalid_params_rpc_err

Constructs an invalid params JSON-RPC error.

pivot_assumption_failed

request_rejected_in_catch_up_mode

request_rejected_too_many_request_error

rpc_err

Constructs a JSON-RPC error, consisting of code, message and optional data.

unimplemented

unknown_block

Type Aliases

EthResult

Result alias