client::rpc::errors

Enum EthApiError

pub enum EthApiError {

    EmptyRawTransactionData,
    FailedToDecodeSignedTransaction,
    InvalidTransactionSignature,
    PoolError(RpcPoolError),
    UnknownBlockNumber,
    UnknownSafeOrFinalizedBlock,
    UnknownBlockOrTxIndex,
    InvalidBlockRange,
    PrevrandaoNotSet,
    ExcessBlobGasNotSet,
    ConflictingFeeFieldsInRequest,
    InvalidTransaction(RpcInvalidTransactionError),
    BothStateAndStateDiffInOverride(Address),
    TransactionNotFound,
    Unsupported(&'static str),
    InvalidParams(String),
    InvalidTracerConfig,
    InvalidRewardPercentiles,
    InternalBlockingTaskError,
    InternalEthError,
    ExecutionTimedOut(Duration),
    InternalJsTracerError(String),
    TransactionInputError(TransactionInputError),
    EvmCustom(String),
    TransactionConversionError,
    Other(String),

}

Errors that can occur when interacting with the eth_ namespace

Variants

EmptyRawTransactionData

When a raw transaction is empty

FailedToDecodeSignedTransaction

When decoding a signed transaction fails

InvalidTransactionSignature

When the transaction signature is invalid

PoolError(RpcPoolError)

Errors related to the transaction pool

UnknownBlockNumber

When an unknown block number is encountered

UnknownSafeOrFinalizedBlock

Thrown when querying for finalized or safe block before the merge transition is finalized, https://github.com/ethereum/execution-apis/blob/6d17705a875e52c26826124c2a8a15ed542aeca2/src/schemas/block.yaml#L109

op-node now checks for either Unknown block OR unknown block: https://github.com/ethereum-optimism/optimism/blob/3b374c292e2b05cc51b52212ba68dd88ffce2a3b/op-service/sources/l2_client.go#L105

TODO(#8045): Temporary, until a version of https://github.com/ethereum-optimism/optimism/pull/10071 is pushed through that doesnโt require this to figure out the EL sync status.

UnknownBlockOrTxIndex

Thrown when an unknown block or transaction index is encountered

InvalidBlockRange

When an invalid block range is provided

PrevrandaoNotSet

An internal error where prevrandao is not set in the evmโs environment

ExcessBlobGasNotSet

excess_blob_gas is not set for Cancun and above

ConflictingFeeFieldsInRequest

Thrown when a call or transaction request (eth_call, eth_estimateGas, eth_sendTransaction) contains conflicting fields (legacy, EIP-1559)

InvalidTransaction(RpcInvalidTransactionError)

Errors related to invalid transactions

BothStateAndStateDiffInOverride(Address)

Thrown when constructing an RPC block from primitive block data fails Thrown when an AccountOverride contains conflicting state and stateDiff fields

TransactionNotFound

Other internal error Error related to signing Thrown when a requested transaction is not found

Unsupported(&'static str)

Some feature is unsupported

InvalidParams(String)

General purpose error for invalid params

InvalidTracerConfig

When the tracer config does not match the tracer

InvalidRewardPercentiles

When the percentile array is invalid

InternalBlockingTaskError

Error thrown when a spawned blocking task failed to deliver an anticipated response.

This only happens if the blocking task panics and is aborted before it can return a response back to the request handler.

InternalEthError

Error thrown when a spawned blocking task failed to deliver an anticipated response

ExecutionTimedOut(Duration)

Error thrown when a (tracing) call exceeds the configured timeout

InternalJsTracerError(String)

Internal Error thrown by the javascript tracer

TransactionInputError(TransactionInputError)

Call Input error when both data and input fields are set and not equal.

EvmCustom(String)

Evm generic purpose error.

TransactionConversionError

Error encountered when converting a transaction type

Other(String)

Error thrown when tracing with a muxTracer fails Any other error