client::rpc

Module types

Modules

address

cfx

pos

pubsub

Pub-Sub types.

transaction_request

Structs

Account

AccountPendingInfo

AccountPendingTransactions

BlameInfo

Block

Bytes

Wrapper structure around vector of bytes.

CfxFeeHistory

CfxRpcLogFilter

CheckBalanceAgainstTransactionResponse

ConsensusGraphBlockExecutionState

ConsensusGraphBlockState

ConsensusGraphStates

This struct maintains some inner state of consensus graph.

EstimateGasAndCollateralResponse

FeeHistory

Header

Block header representation.

Index

Represents usize.

LocalizedBlockTrace

LocalizedTrace

LocalizedTransactionTrace

Log

PoSEconomics

Receipt

RevertTo

RewardInfo

RpcAddress

This is the address type used in Rpc. It deserializes user’s Rpc input, or it prepares the base32 address for Rpc output.

SponsorInfo

StatOnGasLoad

Status

StorageCollateralInfo

SyncGraphBlockState

SyncGraphStates

This struct maintains some inner state of synchronization graph.

TokenSupplyInfo

TraceFilter

Transaction

TransactionRequest

TxPoolPendingNonceRange

TxPoolStatus

TxWithPoolInfo

U64

VoteParamsInfo

Enums

Action

BlockHashOrEpochNumber

BlockTransactions

CfxFilterChanges

Results of the filter_changes RPC.

CfxFilterLog

EpochNumber

Represents rpc api epoch number param.

Origin

RPC request origin

PackedOrExecuted

WrapTransaction

Constants

DEFAULT_CFX_GAS_CALL_REQUEST

The maximum gas limit accepted by most tx pools.

Functions

check_two_rpc_address_network_match