Struct client::Classes

pub struct Classes {
    set: IndexSet<Cow<'static, str>, RandomState>,
}

A set of classes.

The preferred way of creating this is using the classes! macro.

Fields
set: IndexSet<Cow<'static, str>, RandomState>
Implementations

impl Classes

pub fn new() -> Classes

Creates an empty set of classes. (Does not allocate.)

pub fn with_capacity(n: usize) -> Classes

Creates an empty set of classes with capacity for n elements. (Does not allocate if n is zero.)

pub fn push<T>(&mut self, class: T) where
T: Into<Classes>,

Adds a class to a set.

If the provided class has already been added, this method will ignore it.

pub unsafe fn unchecked_push<T>(&mut self, class: T) where
T: Into<Cow<'static, str>>,

Adds a class to a set.

If the provided class has already been added, this method will ignore it.

This method won’t check if there are multiple classes in the input string.

Safety

This function will not split the string into multiple classes. Please do not use it unless you are absolutely certain that the string does not contain any whitespace. Using push() is preferred.

pub fn contains<T>(&self, class: T) -> bool where
T: AsRef<str>,

Check the set contains a class.

pub fn is_empty(&self) -> bool

Check the set is empty.