Struct client::Client

pub struct Client { /* private fields */ }

Instance of a single client for an Aurae consumer.

Implementations

impl Client

pub async fn default() -> Result<Self, ClientError>

pub async fn new(__arg0: AuraeConfig) -> Result<Self, ClientError>

Create a new Client.

Note: A new client is required for every independent execution of this process.

pub async fn new_no_tls(socket: AuraeSocket) -> Result<Self, ClientError>

Create a new Client without TLS, remote server should also expect no TLS.

Note: A new client is required for every independent execution of this process.