Struct client::ClientApp

pub struct ClientApp {
    network: Nela,
    client: Client,
}

The client providing high-level operations that call into the different parts of the core library.

Fields
network: Nelaclient: Client
Implementations

impl ClientApp

fn new(client: Client) -> Self

Login on the backend with username and password

fn display(&self, conversation: String, limit: usize)

fn update(&mut self) -> Result<(), Error>

Pull all new messages from the server and print them out.

fn send_message(&self, msg: String, conversation: String) -> Result<(), Error>

Send a message.

fn register_notification_listener(&self)

Register a callback for push notifications.

fn listener(msg: Message)