Struct client::FocusEvent

#[repr(transparent)]
pub struct FocusEvent {
    obj: UiEvent,
}

The FocusEvent class.

MDN Documentation

This API requires the following crate features to be activated: FocusEvent

Fields
obj: UiEvent
Implementations

impl FocusEvent

pub fn related_target(&self) -> Option<EventTarget>

Getter for the relatedTarget field of this object.

MDN Documentation

This API requires the following crate features to be activated: EventTarget, FocusEvent

impl FocusEvent

pub fn new(type_arg: &str) -> Result<FocusEvent, JsValue>

The new FocusEvent(..) constructor, creating a new instance of FocusEvent.

MDN Documentation

This API requires the following crate features to be activated: FocusEvent