Struct client::test_client::TestClient

pub struct TestClient { /* fields omitted */ }

Client used to test

Methods

impl TestClient

pub fn new(
host: &str,
ac_port: &str,
debug_port: &str,
accounts_file: &str
) -> Self

pub fn get_debug_client(&self) -> &NodeDebugClient

pub fn write_accounts_file(&self, space_delim_strings: &[&str]) -> Result<()>

pub fn create_random_account(
&mut self,
space_delim_strings: &[&str]
) -> Result<AddressAndIndex>

Returns the account index that should be used by user to reference this account

pub fn load_known_accounts(
&mut self,
space_delim_strings: &[&str]
) -> Result<Vec<AddressAndIndex>>

Returns the (account,index) pairs that should be used by user to reference loaded accounts

pub fn print_all_accounts(&self)

pub fn get_balance(&mut self, space_delim_strings: &[&str]) -> Result<u64>

pub fn get_sequence_number(
&mut self,
space_delim_strings: &[&str]
) -> Result<u64>

pub fn mint_coins(
&mut self,
space_delim_strings: &[&str],
is_blocking: bool
) -> Result<()>

Mints coins for the receiver specified

pub fn assert_balance(&mut self, space_delim_strings: &[&str])

pub fn transfer_coins(
&mut self,
space_delim_strings: &[&str],
is_blocking: bool
) -> Result<()>

Transfers coins from sender to receiver

pub fn get_latest_account_state(
&mut self,
space_delim_strings: &[&str]
) -> Result<AccountStateBlob>

Get the latest account state

pub fn get_committed_txn_by_acc_seq(
&mut self,
space_delim_strings: &[&str]
) -> Result<Option<SignedTransaction>>

Get committed txn by account and sequnce number

pub fn get_committed_txn_by_range(
&mut self,
space_delim_strings: &[&str]
) -> Result<Vec<SignedTransaction>>

Get committed txn by account and sequnce number

pub fn get_events_by_access_path(
&mut self,
space_delim_strings: &[&str]
) -> Result<(Vec<EventWithProof>, Option<AccountStateWithProof>)>

Get event by access path