Trait client::CryptoHash

pub trait CryptoHash {
    type Hasher: CryptoHasher;
    fn hash(&self) -> HashValue;
}

A type that implements CryptoHash can be hashed by a cryptographic hash function and produce a HashValue. Each type needs to have its own Hasher type.

Associated Types

type Hasher: CryptoHasher

The associated Hasher type which comes with a unique salt for this type.

Loading content...
Required methods

fn hash(&self) -> HashValue

Hashes the object and produces a HashValue.

Loading content...