Crate client

Modules

logger ğŸ”’

protocol ğŸ”’

Constants

MAX_TASKS ğŸ”’

URL ğŸ”’

Functions

bytes_to_hex_str ğŸ”’

Converts a Vec of Bytes to a String of hexadecimal numbers.

collect_a_task ğŸ”’

Basically generates a task of calculating the hypotenuse of a right triangle, and returns the sides of the triangle as a Vec of Bytes.

get_random_f64 ğŸ”’

Generates a random f64 number.

get_sides ğŸ”’

Generates a pair of random f64 numbers, which will be used as the sides of a right triangle.

main ğŸ”’

process_message ğŸ”’

Process a message, basically check the type of the message and call the corresponding function.

process_response ğŸ”’

Process a message of type MessageType::RESPONSE.

receive_message ğŸ”’

Receives a message from the server and parse it.