Module client::rpc::types

Re-exports

pub use self::cfx::blame_info::BlameInfo;

pub use self::cfx::block::Block;

pub use self::cfx::block::BlockTransactions;

pub use self::cfx::block::Header;

pub use self::cfx::consensus_graph_states::ConsensusGraphStates;

pub use self::cfx::filter::CfxFilterChanges;

pub use self::cfx::filter::CfxFilterLog;

pub use self::cfx::filter::CfxRpcLogFilter;

pub use self::cfx::filter::RevertTo;

pub use self::cfx::log::Log;

pub use self::cfx::pos_economics::PoSEconomics;

pub use self::cfx::pubsub;

pub use self::cfx::receipt::Receipt;

pub use self::cfx::reward_info::RewardInfo;

pub use self::cfx::stat_on_gas_load::StatOnGasLoad;

pub use self::cfx::status::Status;

pub use self::cfx::storage_collateral_info::StorageCollateralInfo;

pub use self::cfx::sync_graph_states::SyncGraphStates;

pub use self::cfx::token_supply_info::TokenSupplyInfo;

pub use self::cfx::transaction::PackedOrExecuted;

pub use self::cfx::transaction::Transaction;

pub use self::cfx::transaction::WrapTransaction;

pub use self::cfx::transaction_request;

pub use self::cfx::transaction_request::CheckBalanceAgainstTransactionResponse;

pub use self::cfx::transaction_request::EstimateGasAndCollateralResponse;

pub use self::cfx::transaction_request::TransactionRequest;

pub use self::cfx::transaction_request::DEFAULT_CFX_GAS_CALL_REQUEST;

pub use self::cfx::tx_pool::AccountPendingInfo;

pub use self::cfx::tx_pool::AccountPendingTransactions;

pub use self::cfx::tx_pool::TxPoolPendingNonceRange;

pub use self::cfx::tx_pool::TxPoolStatus;

pub use self::cfx::tx_pool::TxWithPoolInfo;

pub use self::cfx::vote_params_info::VoteParamsInfo;

pub use self::cfx::Account;

pub use self::cfx::SponsorInfo;

Modules

address

cfx

pos

Structs

Bytes

Wrapper structure around vector of bytes.

CfxFeeHistory

FeeHistory

Index

Represents usize.

LocalizedBlockTrace

LocalizedTrace

LocalizedTransactionTrace

RpcAddress

This is the address type used in Rpc. It deserializes user’s Rpc input, or it prepares the base32 address for Rpc output.

TraceFilter

U64

Enums

Action

BlockHashOrEpochNumber

EpochNumber

Represents rpc api epoch number param.

Origin

RPC request origin

Constants

MAX_GAS_CALL_REQUEST

The MAX_GAS_CALL_REQUEST is used as max value of cfx_call or cfx_estimate’s gas value to prevent call_virtual consumes too much resource. The tx_pool will reject the tx if the gas is larger than half of the block gas limit. which is 30_000_000 before 1559, and 60_000_000 after 1559.

Functions

check_two_rpc_address_network_match