client
Struct Client

pub struct Client {
    pub topology: Option<Config>,
    /* private fields */
}

Fields
topology: Option<Config>
Implementations

impl Client

pub fn new( id: NodeId, controller_send: Sender<DroneEvent>, controller_recv: Receiver<ClientCommand>, packet_recv: Receiver<Packet>, packet_send: HashMap<NodeId, Sender<Packet>>, send_as_if_simulation_controller: Sender<ClientCommand>, ) -> Self

Creates a new Client instance with the provided parameters and starts the web server.

Arguments

id - The node ID for the client.

controller_send - The sender channel to communicate with the simulation controller.

controller_recv - The receiver channel for client commands from the simulation controller.

packet_recv - The receiver channel for incoming packets.

packet_send - A map of packet senders indexed by node ID.

send_as_if_simulation_controller - A sender channel used to simulate sending commands to the controller.